Create networking.py - Add logging warning for #128
